+++ doyto-query-demo/src/test/java/win/doyto/query/user/MockUserRepository.java
@@ -9,7 +9,7 @@ import win.doyto.query.jpa2.AbstractMockRepository;
  * @author f0rb
  * @date 2019-05-15
  */
-public class MockUserRepository extends AbstractMockRepository<UserEntity, Integer, UserQuery> implements UserRepository {
+public class MockUserRepository extends AbstractMockRepository<UserEntity, Long, UserQuery> implements UserRepository {
     public MockUserRepository() {
         super(UserEntity.TABLE);
     }
